testing/mochitest/chrome/test_sanityAddTask.xul

<?xml version="1.0"?>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<window title="Test spawnTawk function"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"/>
  <script type="application/javascript">
    <![CDATA[

      // Check that we can 'add_task' a few times and all tasks run asynchronously before test finishes.

      add_task(function* () {
        var x = yield Promise.resolve(1);
        is(x, 1, "task yields Promise value as expected");
      });

      add_task(function* () {
        var x = yield [Promise.resolve(1), Promise.resolve(2), Promise.resolve(3)];
        is(x.join(""), "123", "task yields Promise value as expected");
      });

      add_task(function* () {
        var x = yield (function* () {
          return 3;
        }());
        is(x, 3, "task yields generator function return value as expected");
      });

    ]]>
  </script>
  <body xmlns="http://www.w3.org/1999/xhtml" >
  </body>
</window>
